refactor(remote): split mobileRemoteServer.ts god-file (DEBT #4)
Extract cohesive seams from the 580-LOC mobileRemoteServer.ts into
sibling modules under electron/remote/: wsProtocol.ts (frame
encode/decode, closeSocket, upgrade response), remoteHttp.ts (URL
parsing, send helpers, constant-time tokenMatches, isRecord),
remoteMessages.ts (handleClientMessage routing), mobilePage.ts (the
served HTML). mobileRemoteServer.ts keeps the server facade + singleton
state (clients set, seqBySid, token).

Verbatim move, no runtime or security change. Auth/pairing token
checks, the constant-time compare, the per-client subscribedSid
cross-session gate, and all WebSocket close-frame flushing are copied
byte-for-byte.
